test_wiener/pre_filtering/noise_reduced_data/requirements.txt
wheel
appdirs
audioread
Cython
decorator
joblib
librosa
numpy
packaging
pyparsing
pysndfx
python-speech-features
resampy
scikit-learn
scipy
six
CFFI
pysoundfile
